scripts: add daemons_run.sh test script
With this script it is possible to run N full daemons on
localhost and had them connect to each other.

They all run with UPnP disabled (as they don't need to talk
to the Internet) and they are all initialized with M
known nodes in order to bootstrap the DHT.

The script can then be used to control the daemons via the
cli tool.

When launched, the script will create a folder named
"daemons_run_tmp". For each of the N daemons, in this folder
you will find:
- a data dir
- a config file (.json)
- a log file containing stdout/err (.log)
- a pid file (.pid)

The script can lastly be used to to stop/kill the running
daemons.

N and M are hardcoded in the script and have a default value
of N=10 and M=3

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>