daemons_run.sh: implement first basic consistency check
the new 'check_blob <blob_hash>' command injects the given
blob on a random daemon which will then announce it in the
DHT. The script will then ask for the peer_list of that
blob to every daemon in the test and compare the output.

Inconsistencies are reported to stdout.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>